pyproject.toml
[project]
name = "kiln"
version = "0.1.0"
description = "Agent runtime for Claude Code — identity-neutral, extensible"
license = "Apache-2.0"
authors = [
    {name = "Kira Selby"},
]
requires-python = ">=3.12"
dependencies = [
    "claude-agent-sdk~=0.1.60",
    "pyyaml~=6.0.3",
    "prompt-toolkit~=3.0.52",
    "markdown-it-py~=4.0.0",
    "openai~=2.31.0",
    "pypdf>=4.0",
    "croniter~=6.0.0",
]

[project.scripts]
kiln = "kiln.cli:main"

[project.optional-dependencies]
discord = [
    "discord.py==2.7.1",
]
dev = [
    "pytest~=9.0.3",
    "pytest-asyncio~=1.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/kiln"]

[tool.pytest.ini_options]
pythonpath = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
